funnel converion view/requirements.txt
Flask==2.3.3
google-cloud-bigquery==3.11.4
google-auth==2.23.4
Werkzeug==2.3.7
Jinja2==3.1.2 